requirements.txt
Django==5.00
pandas==2.1.3
djangorestframework==3.14.0